test settings server

+++ OxApp/console/SettingsTunnelOnServer.php
@@ -47,24 +47,25 @@ class SettingsTunnelOnServer extends Command
             Tunnels::where(['id' => $tunnel->id])->update([
                 'status' => 3,
             ]);
-            if ($tunnel->v6route !== '[1500]') {
+            if ($tunnel->v6route !== '[1500]' && $tunnel->remoteIp !== '[1500]') {
                 $server = Servers::find(['ip' => $tunnel->serverIp])->rows[0];
                 print_r($server);
                 $connection = ssh2_connect($server->ip, 22);
                 var_dump(ssh2_auth_password($connection, 'root', $server->password));
     
                 ssh2_exec($connection, 'pkill 3proxy');
-                ssh2_exec($connection, 'ulimit -n 600000');
-                ssh2_exec($connection, 'ulimit -u 600000');
-                ssh2_exec($connection, 'ip link delete he-ipv6');
-                ssh2_exec($connection, 'ifconfig he-ipv6 down');
-                ssh2_exec($connection, 'ip -6 route del default');
-                ssh2_exec($connection, 'modprobe ipv6');
-                ssh2_exec($connection,
-                    'ip tunnel add he-ipv6 mode sit remote ' . $tunnel->remoteIp . ' local ' . $tunnel->serverIp . ' ttl 255');
-                ssh2_exec($connection, 'ip link set he-ipv6 up');
-                ssh2_exec($connection, 'ip addr add ' . $tunnel->v6route . ' dev he-ipv6');
-                ssh2_exec($connection, 'ip route add ::/0 dev he-ipv6');
+                ssh2_exec($connection, 'sh 1.sh '. $tunnel->remoteIp.' '. $tunnel->v6route);
+//                ssh2_exec($connection, 'ulimit -n 600000');
+//                ssh2_exec($connection, 'ulimit -u 600000');
+//                ssh2_exec($connection, 'ip link delete he-ipv6');
+//                ssh2_exec($connection, 'ifconfig he-ipv6 down');
+//                ssh2_exec($connection, 'ip -6 route del default');
+//                ssh2_exec($connection, 'modprobe ipv6');
+//                ssh2_exec($connection,
+//                    'ip tunnel add he-ipv6 mode sit remote ' . $tunnel->remoteIp . ' local ' . $tunnel->serverIp . ' ttl 255');
+//                ssh2_exec($connection, 'ip link set he-ipv6 up');
+//                ssh2_exec($connection, 'ip addr add ' . $tunnel->v6route . ' dev he-ipv6');
+//                ssh2_exec($connection, 'ip route add ::/0 dev he-ipv6');
     
                 $name = '48sub';
                 $exc = './fastProxy.sh ' . $tunnel->$name;
